requirements.txt
streamlit>=1.28.0
sec-edgar-downloader>=5.0.0
google-generativeai>=0.8.0
beautifulsoup4>=4.12.0
requests>=2.31.0
pandas>=2.0.0
lxml>=4.9.0
python-dotenv>=1.0.0
yfinance>=0.2.40
